backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
groq>=0.4.2
networkx>=3.2.1
pydantic>=2.6.4
supabase>=2.4.1
httpx>=0.27.0
python-dotenv>=1.0.1